+++ src/components/zshrc/nvm_switcher.sh
@@ -1,0 +1,17 @@
+#!/bin/bash
+
+export NVM_DIR="$HOME/.nvm"
+[ -s "/usr/local/opt/nvm/nvm.sh" ] && . "/usr/local/opt/nvm/nvm.sh"  # This loads nvm
+[ -s "/usr/local/opt/nvm/etc/bash_completion.d/nvm" ] && . "/usr/local/opt/nvm/etc/bash_completion.d/nvm"  # This loads nvm bash_completion
+
+# Automatically flip node version on CD if there a .nvmrc file in the repo
+autoload -U add-zsh-hook
+load-nvmrc() {
+    local nvmrc_path="$(nvm_find_nvmrc)"
+    if [ -n "$nvmrc_path" ]; then
+        nvm install
+    else
+        nvm use default
+    fi
+}
+add-zsh-hook chpwd load-nvmrc