+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2020-06-12
+### Added
+- Install Docker, Kubernetes, and Minikube
+- Add NVM stuff to .zshrc
+- Install Intellij IDEA
+- Install Postman
+- Install Spotify
+
 ## [0.1.0] - 2020-06-12
 ### Added
 - Install Homebrew